requirements.txt
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-API==3.0.post1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.21.0
SQLAlchemy==1.4.41
Werkzeug==2.2.2
zipp==3.8.1
cohere
umap-learn 
altair 
annoy 
datasets 
tqdm